+++ hub20/api/celery.py
@@ -1,58 +1,14 @@
-import json
 import os
-import uuid
 from datetime import timedelta
-from decimal import Decimal
 
 from celery import Celery
 from celery.schedules import crontab
 from django.conf import settings
-from hexbytes import HexBytes
 from kombu.serialization import register
-from web3.datastructures import AttributeDict
+
+from hub20.apps.blockchain.celery import web3_deserializer, web3_serializer
 
 NODE_HEALTH_CHECK_INTERVAL = 15
-
-
-class Web3Encoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, AttributeDict):
-            return {
-                "__type__": "attrdict",
-                "value": json.dumps({k: v for k, v in obj.items()}, cls=self.__class__),
-            }
-        elif isinstance(obj, HexBytes):
-            return {"__type__": "hexbytes", "value": obj.hex()}
-        elif isinstance(obj, uuid.UUID):
-            return {"__type__": "uuid", "value": str(obj)}
-        elif isinstance(obj, Decimal):
-            return {"__type__": "decimal", "value": str(obj)}
-        else:
-            return json.JSONEncoder.default(self, obj)
-
-
-def web3_decoder(obj):
-    try:
-        if obj["__type__"] == "attrdict":
-            return AttributeDict(json.loads(obj["value"], object_hook=web3_decoder))
-        elif obj["__type__"] == "hexbytes":
-            return HexBytes(obj["value"])
-        elif obj["__type__"] == "uuid":
-            return uuid.UUID(obj["value"])
-        elif obj["__type__"] == "decimal":
-            return Decimal(obj["value"])
-        else:
-            return obj
-    except KeyError:
-        return obj
-
-
-def web3_serializer(obj):
-    return json.dumps(obj, cls=Web3Encoder)
-
-
-def web3_deserializer(obj):
-    return json.loads(obj, object_hook=web3_decoder)
 
 
 register(