+++ hub20/apps/core/api.py
@@ -9,6 +9,15 @@ from .consumers import CheckoutConsumer, SessionEventsConsumer
 
 app_name = "hub20"
 
+
+class UUIDRouter(SimpleRouter):
+    lookup_value_regex = "[0-9a-f-]{36}"
+
+
+class TransferRouter(UUIDRouter):
+    lookup_field = "reference"
+
+
 token_router = TokenRouter(trailing_slash=False)
 token_router.register(r"tokens", views.TokenBrowserViewSet, basename="token")
 token_router.register(r"my/tokens", views.UserTokenViewSet, basename="user-token")
@@ -25,6 +34,12 @@ router.register(
 )
 router.register("my/stores", views.UserStoreViewSet, basename="user-store")
 router.register("my/tokenlists", UserTokenListViewSet, basename="user-tokenlist")
+router.register("my/deposits", views.DepositViewSet, basename="user-deposit")
+
+transfer_router = TransferRouter(trailing_slash=False)
+transfer_router.register("my/transfers", views.TransferViewSet, basename="user-transfer")
+transfer_router.register("my/withdrawals", views.WithdrawalViewSet, basename="user-withdrawal")
+
 
 urlpatterns = (
     [
@@ -36,21 +51,18 @@ urlpatterns = (
             views.TokenBalanceView.as_view(),
             name="balance-detail",
         ),
-        path("deposits", views.DepositListView.as_view(), name="deposit-list"),
-        path("deposit/<uuid:pk>", views.DepositView.as_view(), name="deposit-detail"),
         path("payment/orders", views.PaymentOrderListView.as_view(), name="payment-order-list"),
         path(
             "payment/order/<uuid:pk>",
             views.PaymentOrderView.as_view(),
             name="payment-order-detail",
         ),
-        path("transfers", views.TransferListView.as_view(), name="transfer-list"),
-        path("transfers/transfer/<int:pk>", views.TransferView.as_view(), name="transfer-detail"),
         path("accounting/report", views.AccountingReportView.as_view(), name="accounting-report"),
         path("my/settings", views.UserPreferencesView.as_view(), name="user-preferences"),
     ]
     + router.urls
     + token_router.urls
+    + transfer_router.urls
 )