+++ hub20/apps/ethereum_money/typing.py
@@ -4,5 +4,5 @@ from typing import TypeVar, Union
 Wei = int
 TokenAmount = Decimal
 
-EthereumClient_T = TypeVar("EthereumClient_T")
+Web3Client_T = TypeVar("Web3Client_T")
 TokenAmount_T = Union[int, float, Decimal, TokenAmount, Wei]