+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+ - Refactored transfer model, broken from single Transfer class into
+   Internal/Blockchain/Raiden transfers, and corresponding
+   execution/confirmation models
+
 ## [0.4.0] - 2022-01-31
 
 ### Added